feat(comms): messages above BUFFER_SIZE can now be sent over client

+++ src/comms/message_reset.c
@@ -6,15 +6,16 @@
 /*   By: emcnab <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/01 13:11:14 by emcnab            #+#    #+#             */
-/*   Updated: 2023/03/01 13:12:40 by emcnab           ###   ########.fr       */
+/*   Updated: 2023/03/02 13:56:54 by emcnab           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "message_reset.h"
+#include <stddef.h>
 #include <stdint.h>
 
 void	message_reset(t_s_message *message)
 {
-	message->mask = (int8_t)0x80;
+	message->mask = (size_t)0x80;
 	message->bit_count = 0;
 }
